Send errors async and shutdown afterwards (#12)
* Send errors async and shutdown afterwards

In some cases the program exits before the error can be sent. This
ensures the error is sent asynchronously and shuts down the executor
when the thread exits.

* add to .gitignore

* Test fix attempts

* Switch to runAsync since we don't need a return

* Rename to MockSyncSender for clarity

This class is just a helper so we can inspect the notices that would
have been sent and to avoid actual sending behavior in the tests.
Renaming it makes it more clear that this is not a test.